+++ src/app/MyPage/page.tsx
@@ -1,8 +1,11 @@
-import React from 'react';
+'use client';
+
+import Header from '@/components/Header';
 
 function MyPage() {
 	return (
 		<div>
+			<Header />
 			<div>MyPage</div>
 		</div>
 	);
